Apply com.gradle.plugin-publish last
Depending on the artifacts in the archives configuration, the
`com.gradle.plugin-publish` plugin creates tasks to generate JARs for
Javadoc, Groovydoc, and sources. To do so, it registers an
`afterEvaluate` hook and checks the size of the artifacts in the
archives configuration. Since this plugin also creates tasks for such
JARs in an `afterEvalutate` hook, we need to ensure that we run before
the `com.gradle.plugin-publish` plugin.