Don't add jar task to archives twice
Prior to this commit, the `jar` task was unconditionally added to the
artifacts of the archives configuration even though it was already added
by the `java` plugin. The resulting duplicate was silently ignored by
Gradle until now but causes an exception when executing `signArchives`
in Gradle 5.1.

In addition, this commit resolves the TODO when the `maven-publish`
plugin is used since the `jar` is already added to the published
artifacts by the used `components.java`.

Fixes #38.